% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppcc.R
\docType{package}
\name{ppcc-package}
\alias{ppcc}
\alias{ppcc-package}
\title{Goodness-of-Fit Tests using the
Probability Plot Correlation Coefficient}
\description{
The function \code{\link{ppccTest}} performs
the Probability Plot Correlation Coefficient test
for various continuous distribution functions.
}
